NiceReferenceImageDBCreator/requirements.txt
azure-functions
numpy
requests